feat(auth): refactor authentication provider and session management
